requirements.txt
ccxt==2.*
python-dotenv
pandas
numpy
requests
scikit-learn
vectorbt
ta
aiohttp
python-telegram-bot
sqlalchemy